# Add metadata exchange as native code