--- a/pilot/pkg/networking/core/v1alpha3/gateway.go
+++ b/pilot/pkg/networking/core/v1alpha3/gateway.go
@@ -107,7 +107,7 @@ func (configgen *ConfigGeneratorImpl) buildGatewayListeners(builder *ListenerBui
 				bind:       bind,
 				port:       &model.Port{Port: int(port.Number)},
 				bindToPort: true,
-				class:      ListenerClassGateway,
+				class:      istionetworking.ListenerClassGateway,
 				transport:  transport,
 			}
 			lname := getListenerName(bind, int(port.Number), transport)