--- a/pilot/pkg/xds/filters/filters.go
+++ b/pilot/pkg/xds/filters/filters.go
@@ -16,6 +16,7 @@ package filters
 
 import (
 	udpa "github.com/cncf/udpa/go/udpa/type/v1"
+	cluster "github.com/envoyproxy/go-control-plane/envoy/config/cluster/v3"
 	core "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
 	listener "github.com/envoyproxy/go-control-plane/envoy/config/listener/v3"
 	cors "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/cors/v3"
@@ -23,16 +24,15 @@ import (
 	grpcstats "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/grpc_stats/v3"
 	grpcweb "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/grpc_web/v3"
 	router "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/router/v3"
-	wasm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/wasm/v3"
+	httpwasm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/wasm/v3"
 	httpinspector "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/listener/http_inspector/v3"
 	originaldst "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/listener/original_dst/v3"
 	originalsrc "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/listener/original_src/v3"
 	tlsinspector "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/listener/tls_inspector/v3"
 	hcm "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
-	v3 "github.com/envoyproxy/go-control-plane/envoy/extensions/wasm/v3"
+	wasm "github.com/envoyproxy/go-control-plane/envoy/extensions/wasm/v3"
 	"github.com/envoyproxy/go-control-plane/pkg/wellknown"
 	protobuf "github.com/gogo/protobuf/types"
-	"google.golang.org/protobuf/types/known/anypb"
 	"google.golang.org/protobuf/types/known/structpb"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 
@@ -138,20 +138,26 @@ var (
 		},
 	}
 
-	TCPMx = &listener.Filter{
-		Name: MxFilterName,
-		// TODO: we need to publish this tcp proto: https://github.com/istio/proxy/blob/master/src/envoy/tcp/metadata_exchange/config/metadata_exchange.proto
-		// somewhere as go code
-		ConfigType: &listener.Filter_TypedConfig{TypedConfig: util.MessageToAny(&udpa.TypedStruct{
-			TypeUrl: "type.googleapis.com/envoy.tcp.metadataexchange.config.MetadataExchange",
-			Value: &structpb.Struct{
-				Fields: map[string]*structpb.Value{
-					"protocol": {
-						Kind: &structpb.Value_StringValue{StringValue: "istio-peer-exchange"},
-					},
+	tcpMx = util.MessageToAny(&udpa.TypedStruct{
+		// TODO(https://github.com/istio/istio/issues/35210) make this in API
+		TypeUrl: "type.googleapis.com/envoy.tcp.metadataexchange.config.MetadataExchange",
+		Value: &structpb.Struct{
+			Fields: map[string]*structpb.Value{
+				"protocol": {
+					Kind: &structpb.Value_StringValue{StringValue: "istio-peer-exchange"},
 				},
 			},
-		})},
+		},
+	})
+
+	TCPListenerMx = &listener.Filter{
+		Name:       MxFilterName,
+		ConfigType: &listener.Filter_TypedConfig{TypedConfig: tcpMx},
+	}
+
+	TCPClusterMx = &cluster.Filter{
+		Name:        MxFilterName,
+		TypedConfig: tcpMx,
 	}
 
 	HTTPMx = buildHTTPMxFilter()
@@ -185,51 +191,50 @@ var (
 )
 
 func buildHTTPMxFilter() *hcm.HttpFilter {
-	var vmConfig *v3.PluginConfig_VmConfig
+	httpMxConfigProto := &httpwasm.Wasm{
+		Config: &wasm.PluginConfig{
+			Vm:            constructVMConfig("/etc/istio/extensions/metadata-exchange-filter.compiled.wasm", "envoy.wasm.metadata_exchange"),
+			Configuration: util.MessageToAny(&protobuf.StringValue{Value: "{}"}),
+		},
+	}
 
+	return &hcm.HttpFilter{
+		Name:       MxFilterName,
+		ConfigType: &hcm.HttpFilter_TypedConfig{TypedConfig: util.MessageToAny(httpMxConfigProto)},
+	}
+}
+
+// constructVMConfig constructs a VM config. If WASM is enabled, the wasm plugin at filename will be used.
+// If not, the builtin (null vm) extension, name, will be used.
+func constructVMConfig(filename, name string) *wasm.PluginConfig_VmConfig {
+	var vmConfig *wasm.PluginConfig_VmConfig
 	if features.EnableWasmTelemetry {
-		vmConfig = &v3.PluginConfig_VmConfig{
-			VmConfig: &v3.VmConfig{
+		vmConfig = &wasm.PluginConfig_VmConfig{
+			VmConfig: &wasm.VmConfig{
 				Runtime:          "envoy.wasm.runtime.v8",
 				AllowPrecompiled: true,
 				Code: &core.AsyncDataSource{Specifier: &core.AsyncDataSource_Local{
 					Local: &core.DataSource{
 						Specifier: &core.DataSource_Filename{
-							Filename: "/etc/istio/extensions/metadata-exchange-filter.compiled.wasm",
+							Filename: filename,
 						},
 					},
 				}},
 			},
 		}
 	} else {
-		vmConfig = &v3.PluginConfig_VmConfig{
-			VmConfig: &v3.VmConfig{
+		vmConfig = &wasm.PluginConfig_VmConfig{
+			VmConfig: &wasm.VmConfig{
 				Runtime: "envoy.wasm.runtime.null",
 				Code: &core.AsyncDataSource{Specifier: &core.AsyncDataSource_Local{
 					Local: &core.DataSource{
 						Specifier: &core.DataSource_InlineString{
-							InlineString: "envoy.wasm.metadata_exchange",
+							InlineString: name,
 						},
 					},
 				}},
 			},
 		}
 	}
-
-	httpMxConfigProto := &wasm.Wasm{
-		Config: &v3.PluginConfig{
-			Vm:            vmConfig,
-			Configuration: util.MessageToAny(&protobuf.StringValue{Value: "{}\n"}),
-		},
-	}
-
-	typed, err := anypb.New(httpMxConfigProto)
-	if err != nil {
-		return nil
-	}
-
-	return &hcm.HttpFilter{
-		Name:       MxFilterName,
-		ConfigType: &hcm.HttpFilter_TypedConfig{TypedConfig: typed},
-	}
+	return vmConfig
 }