… manifests/charts/istio-control/istio-discovery/files/gen-istio.yaml
… manifests/charts/istio-control/istio-discovery/templates/telemetryv2_1.10.yaml
… manifests/charts/istio-control/istio-discovery/templates/telemetryv2_1.11.yaml
… manifests/charts/istio-control/istio-discovery/templates/telemetryv2_1.12.yaml
… pilot/pkg/networking/core/v1alpha3/listener_test.go