Benchmark:
```
name                            old time/op        new time/op        delta
ClusterGeneration/telemetry-6          168µs ± 1%         155µs ± 2%  -7.94%  (p=0.029 n=4+4)
ListenerGeneration/telemetry-6         512µs ± 3%         501µs ± 1%  -1.98%  (p=0.114 n=4+4)

name                            old kb/msg         new kb/msg         delta
ClusterGeneration/telemetry-6           5.53 ± 0%          5.34 ± 0%  -3.42%  (p=0.029 n=4+4)
ListenerGeneration/telemetry-6          32.9 ± 0%          31.9 ± 0%  -3.16%  (p=0.029 n=4+4)

name                            old resources/msg  new resources/msg  delta
ClusterGeneration/telemetry-6           14.0 ± 0%          14.0 ± 0%    ~     (all equal)
ListenerGeneration/telemetry-6          6.00 ± 0%          6.00 ± 0%    ~     (all equal)

name                            old alloc/op       new alloc/op       delta
ClusterGeneration/telemetry-6         76.1kB ± 0%        70.8kB ± 0%  -6.96%  (p=0.029 n=4+4)
ListenerGeneration/telemetry-6         254kB ± 0%         251kB ± 0%  -1.20%  (p=0.029 n=4+4)

name                            old allocs/op      new allocs/op      delta
ClusterGeneration/telemetry-6          1.19k ± 0%         1.10k ± 0%  -7.24%  (p=0.029 n=4+4)
ListenerGeneration/telemetry-6         3.15k ± 0%         3.12k ± 0%  -1.14%  (p=0.029 n=4+4)
```

Broken out of https://github.com/istio/istio/pull/33583

This moves the MX code from EnvoyFilter to native code. This is an important part of the Telemetry API implementation.

In a followup, the remaining filters will be converted as well.

The MX filter is now unconditionally enabled, similar to other Istio features. This is because we need the MX filter on both the client *and* server in order for telemetry to work. Because we cannot know the entirety of the Telemetry config for client/servers (especially cross cluster), there is no reliable way to detect if its safe to elide MX filter. For users that have extreme high performance use cases and do not use Istio telemetry, nor any other filters depending on MX, there is an opt out flag left in Istiod. We expect this to be the minority use case (@ramaraochavali :slightly_smiling_face:).