@ramaraochavali 2021-09-16
Should we also fold PILOT_ENABLE_TELEMETRY_LABEL and PILOT_ENDPOINT_TELEMETRY_LABEL in this config? Or MetadataExchange can be independently enabled?