@howardjohn 2021-09-16
> Should we also fold PILOT_ENABLE_TELEMETRY_LABEL and PILOT_ENDPOINT_TELEMETRY_LABEL in this config? Or MetadataExchange can be independently enabled?

Like have 1 env var to control all of them or something? I would be ok with that I think. You can technically use MX without Istio telemetry but that is so obscure I don't think we care too much (?)