@bianpengyuan 2021-09-16
> Should we also fold PILOT_ENABLE_TELEMETRY_LABEL and PILOT_ENDPOINT_TELEMETRY_LABEL in this config? Or MetadataExchange can be independently enabled?

I think there can be case where user uses MX with envoy access log service. Better have separated control if it is not too much of burden. Setting TELEMETRY_LABEL to false removes MX feels a bit too magic to me.