@ramaraochavali 2021-09-17
Ok. I am fine with leaving them as is. I was thinking we use the new flag
that was added here to control telemetry labels also I.e if metadata
exchange is enabled we add labels.

On Thu, 16 Sep 2021 at 10:00 PM, Pengyuan Bian ***@***.***>
wrote:

> Should we also fold PILOT_ENABLE_TELEMETRY_LABEL and
> PILOT_ENDPOINT_TELEMETRY_LABEL in this config? Or MetadataExchange can be
> independently enabled?
>
> I think there can be case where user uses MX with envoy access log
> service. Better have separated control if it is not too much of burden.
> Setting TELEMETRY_LABEL to false removes MX feels a bit too magic to me.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/istio/istio/pull/35212#issuecomment-921053029>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AEDINKCWXRKONKB7YGA4IVTUCILSPANCNFSM5EDIZIFA>
> .
> Triage notifications on the go with GitHub Mobile for iOS
> <https://apps.apple.com/app/apple-store/id1477376905?ct=notification-email&mt=8&pt=524675>
> or Android
> <https://play.google.com/store/apps/details?id=com.github.android&referrer=utm_campaign%3Dnotification-email%26utm_medium%3Demail%26utm_source%3Dgithub>.
>
>